endpoint: move restore functions to `restore.go`
Signed-off by: Ian Vernon <ian@cilium.io>
